fix: move custom error class to avoid importing unsupported syntax in tests
