+++ src/indexer/utils.ts
@@ -8,14 +8,3 @@ export async function importAbi(path: string): Promise<ContractInterface> {
 
   return importResult.default;
 }
-
-export class UnknownTokenError extends Error {
-  public constructor(
-    public address: string,
-    public chainId: number,
-    message?: string
-  ) {
-    super(message ?? `Token ${address} not configured for chain ${chainId}`);
-    this.name = new.target.name;
-  }
-}
